requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==23.0.0
idna==3.10
openrouteservice==2.3.3
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.0.1
python-http-client==3.3.7
requests==2.32.3
sendgrid==3.6.5
sendgrid-django==4.2.0
sqlparse==0.5.3
stripe==11.5.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
